bmv_hybrid_clean_v3/requirements-lite.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
yfinance>=0.2.40
ta>=0.11.0
PyYAML>=6.0.1
pytz>=2024.1
python-dateutil>=2.9.0
requests>=2.31.0
joblib>=1.3.0
